Export MemoryCookieStore. Fix #39.
